extract statewrapper
